Move parseIncomingMessage to tests/helpers/comparison folder
