+++ packages/core/comparison/lib/comparePhase1.ts
@@ -15,9 +15,9 @@ import getMessageType from "../../phase1/lib/getMessageType"
 import phase1Handler from "../../phase1/phase1"
 import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "../../tests/helpers/comparison/getPncQueryTimeFromAho"
+import parseIncomingMessage from "../../tests/helpers/comparison/parseIncomingMessage"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import isIntentionalDifference from "./isIntentionalDifference"
-import parseIncomingMessage from "./parseIncomingMessage"
 import { sortExceptions } from "./sortExceptions"
 import { sortTriggers } from "./sortTriggers"
 import { matchingExceptions } from "./summariseMatching"
