+++ packages/core/comparison/lib/comparePhase2.ts
@@ -12,9 +12,9 @@ import { parsePncUpdateDataSetXml } from "../../lib/parse/parsePncUpdateDataSetX
 import serialiseToXml from "../../lib/serialise/pncUpdateDatasetXml/serialiseToXml"
 import getMessageType from "../../phase1/lib/getMessageType"
 import phase2Handler from "../../phase2/phase2"
+import parseIncomingMessage from "../../tests/helpers/comparison/parseIncomingMessage"
 import extractAuditLogEventCodes from "./extractAuditLogEventCodes"
 import isIntentionalDifference from "./isIntentionalDifference"
-import parseIncomingMessage from "./parseIncomingMessage"
 import { sortExceptions } from "./sortExceptions"
 import { sortTriggers } from "./sortTriggers"
 import { xmlOutputDiff, xmlOutputMatches } from "./xmlOutputComparison"
