+++ packages/core/comparison/lib/comparePhase3.ts
@@ -19,10 +19,10 @@ import { isPncLockError } from "../../phase3/exceptions/generatePncUpdateExcepti
 import { MAXIMUM_PNC_LOCK_ERROR_RETRIES } from "../../phase3/lib/updatePnc"
 import phase3Handler from "../../phase3/phase3"
 import getPncOperationsFromPncUpdateDataset from "../../phase3/tests/helpers/getPncOperationsFromPncUpdateDataset"
+import parseIncomingMessage from "../../tests/helpers/comparison/parseIncomingMessage"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
 import extractAuditLogEventCodes from "./extractAuditLogEventCodes"
-import parseIncomingMessage from "./parseIncomingMessage"
 import { sortExceptions } from "./sortExceptions"
 import { sortTriggers } from "./sortTriggers"
 import { xmlOutputDiff, xmlOutputMatches } from "./xmlOutputComparison"
