+++ packages/core/comparison/lib/comparePncMatching.ts
@@ -10,8 +10,8 @@ import { parseAhoXml } from "../../lib/parse/parseAhoXml"
 import CorePhase1 from "../../phase1/phase1"
 import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/generateMockPncQueryResultFromAho"
 import getPncQueryTimeFromAho from "../../tests/helpers/comparison/getPncQueryTimeFromAho"
+import parseIncomingMessage from "../../tests/helpers/comparison/parseIncomingMessage"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import parseIncomingMessage from "./parseIncomingMessage"
 import summariseMatching from "./summariseMatching"
 
 type CompareOptions = {
